[FE: feat]: 채팅방 생성 API 연결중 본문: 요청에 /spring빠져서 보냄 뭐지...? 현재 위치 행정구역으로 바꾸는 거는 일단 유기

+++ fe/D104/app/src/main/java/com/d104/pnt/data/remote/model/response/ChatResponse.kt
@@ -8,22 +8,22 @@ data class ChatCreateRequest(
     @SerializedName("description")
     val description: String,
     @SerializedName("regionCode")
-    val regionCode: Int,
+    val regionCode: Long,
     @SerializedName("maxMember")
     val maxMember: Int
 )
 
 data class ChatCreateResponse(
     @SerializedName("chatRoomId")
-    val chatRoomId: Int,
+    val chatRoomId: Long,
     @SerializedName("ownerId")
-    val ownerId: Int,
+    val ownerId: Long,
     @SerializedName("title")
     val title: String,
     @SerializedName("description")
     val description: String,
     @SerializedName("regionCode")
-    val regionCode: Int,
+    val regionCode: Long,
     @SerializedName("maxMember")
     val maxMember: Int,
     @SerializedName("currentMember")
